The "end" event that was supposed to be emitted to fix a core bug in NodeJS wasn't fired because it wasn't emitted on the response object.
I encountered this problem when trying to fetch https://profiles.google.com/VoxPelli through JSDom. I tracked it down to being that the response object in this file only emitted "data"-events and no "end"-events. The change here fixed that.

(I hope the bonus spelling correction of "because" is okay as well :) )