Request.multipart no longer crashes when header 'Content-type' is present
Request.multipart used to crash when options contained a header 'Content-type' which is not purely lowercase. That is because hasHeader is case insensitive and returns true and then the split method gets applied to undefined and crashes the application. One could argue that headers should be spelled using only lowercase. However, mixed case headers used to work with a previous version of request, thus I suggest to incorporate this patch to ensure that legacy code does not suddenly crash after updating request.