[PATCH v4] Clarify gzip handling in the README
Make it more explicit in the documentation about which data is
unmodified and which is automatically decoded when using the `gzip`
option (data passed through request is decoded while the response object
is unmodified and thus its data may be compressed).  Add an example to
README.md to clarify.

Changes since v3:
 - Rebased against current master.

Changes since v2:
 - Make note of gzip example in examples section to alleviate confusion.

Changes since v1:
 - Reworded to refer to streams generically rather than the 'data' event
   specifically.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>