Fix multiple redirects and `self.followRedirect`
Currently when a redirect occurs, the `self.init` function is being run. However, since it’s already been run on the initial execution, the original value for `self.followRedirect` has been lost. This means the custom function for `self.allowRedirect` will only ever work on the first redirect and not subsequent redirects.